refactoring: move some entities into /entity/detection package
