+++ app/lib/util/constants.dart
@@ -6,4 +6,7 @@ class Constants {
 
   static const lightBackgroundColor = 0xffffffff;
   static var lightIconsColor = Colors.black;
+
+// static const darkBackgroundColor = 0x00080808;
+// static var darkIconsColor = Colors.white;
 }
